Made felix/frenetic frankencode
I've sewn together the netkat-auatomata/felix code into the frenetic
code in a very rough around the edges kind of way. I've made it so that
the frenetic shell now understands a felix command. Next up, I'll throw
in a verification command and then start working on cleaning up the
merge and removing dependencies and whatnot.