Started tidying up Frenetic_Decide_Util.
The code in Frenetic_Decide_Util is quite old and was starting to show
its age. There was a lot of stuff in the mli that wasn't being used at
all; some of the modules, like Worklist, really ought to have been moved
around; etc.

I started tidying up the code, though there is still lots left to do.
Most of the types should be moved in the Ast file. Also, they don't use
Core!