Expand scope of fabric stream bucketing
Fabric streams are grouped into buckets, one per policy
stream. Previously this bucketing would only be done based on physical
proximity (if fabric stream endpoints are same as, or adjacent to policy
stream endpoints), meaning that code generation could fail if the alphas
of the fabric and policy streams could not be matched. Now, bucketing
takes account of whether the fabric stream alphas can be satisfied as
well. Thus, each fabric stream in a bucket is fully capable of carrying
the policy stream, removing the possibility of failure at the code
generation stage.