Add type and supporting module for analysis form of NetKAT programs.
The analysis form of is a NetKAT program is a policy, guarded by the
ingresses, sequenced and iterated with the topology and guarded with the
egresses, i.e. in;(p.t)*;p;out .

For most cases, this is the correct form to passed to the local compiler
to produce an FDD, and then to generate dyads to use for synthesis. This
commit provides a type and related module to represent the analysis form
so that in general only the assembled analysis form is passed to the
compiler.

There is an "escape hatch" function to produce dyads from a raw policy.